refactoring, plus some fix on how to pars timestamp field
