+++ druid/src/test/java/org/apache/calcite/test/DruidAdapterIT.java
@@ -1503,8 +1503,8 @@ public class DruidAdapterIT {
         + "'dimension':'__time','outputName':'extract',"
         + "'extractionFn':{'type':'timeFormat','format':'yyyy-MM-dd";
     sql(sql)
-        .queryContains(druidChecker(druidQuery))
-        .returnsUnordered("product_id=1016; time=1997-01-02 00:00:00");
+        .returnsUnordered("product_id=1016; time=1997-01-02 00:00:00")
+        .queryContains(druidChecker(druidQuery));
   }
 
   @Test public void testPushAggregateOnTimeWithExtractYear() {
@@ -1977,7 +1977,13 @@ public class DruidAdapterIT {
         "C=6662; S=20388; EXPR$2=1997-09-01 00:00:00",
         "C=6588; S=20179; EXPR$2=1997-04-01 00:00:00",
         "C=6478; S=19958; EXPR$2=1997-10-01 00:00:00")
-        .queryContains(druidChecker("'queryType':'timeseries'"));
+        .queryContains(druidChecker("'queryType':'timeseries'"))
+        .explainContains("PLAN=EnumerableInterpreter\n"
+            + "  BindableSort(sort0=[$1], dir0=[DESC])\n"
+            + "    BindableProject(C=[$1], S=[$2], EXPR$2=[CAST($0):TIMESTAMP(0) NOT NULL])\n"
+            + "      DruidQuery(table=[[foodmart, foodmart]], "
+            + "intervals=[[1900-01-09T00:00:00.000Z/2992-01-10T00:00:00.000Z]], projects=[[FLOOR($0, "
+            + "FLAG(MONTH)), $89]], groups=[{0}], aggs=[[COUNT(), SUM($1)]])");
   }
 
   @Test public void testNumericOrderingOfOrderByOperatorFullTime() {
@@ -2322,11 +2328,11 @@ public class DruidAdapterIT {
         + "    DruidQuery(table=[[foodmart, foodmart]], "
         + "intervals=[[1900-01-09T00:00:00.000Z/2992-01-10T00:00:00.000Z]], filter=[AND(=(";
     sql(sql, FOODMART)
-        .explainContains(plan)
-        .queryContains(druidChecker(druidQuery))
         .returnsOrdered("store_state=CA; brand_name=Bird Call; A=34.364599999999996",
             "store_state=OR; brand_name=Bird Call; A=39.16359999999999",
-            "store_state=WA; brand_name=Bird Call; A=53.742500000000014");
+            "store_state=WA; brand_name=Bird Call; A=53.742500000000014")
+        .explainContains(plan)
+        .queryContains(druidChecker(druidQuery));
   }
 
   @Test public void testSingleAverageFunction() {
@@ -2385,11 +2391,11 @@ public class DruidAdapterIT {
         + "aggs=[[SUM($90), SUM($91)]], post_projects=[[$0, +($1, 100), -(+($1, 100), $2)]], "
         + "sort0=[1], dir0=[DESC]";
     sql(sqlQuery, FOODMART)
-        .explainContains(plan)
-        .queryContains(druidChecker(postAggString))
         .returnsOrdered("store_state=WA; A=263893.2200000001; C=158568.91210000002",
             "store_state=CA; A=159267.83999999994; C=95737.41489999992",
-            "store_state=OR; A=142377.06999999992; C=85604.56939999988");
+            "store_state=OR; A=142377.06999999992; C=85604.56939999988")
+        .explainContains(plan)
+        .queryContains(druidChecker(postAggString));
   }
 
   @Test public void testDivideByZeroDoubleTypeInfinity() {
@@ -3896,12 +3902,12 @@ public class DruidAdapterIT {
         + "CAST(('1997' || '-01' || '-01') AS DATE) = CAST(\"timestamp\" AS DATE) "
         + "GROUP BY \"timestamp\"";
     sql(sql, FOODMART)
+        .returnsOrdered("EXPR$0=1997-01-01 00:00:00; EXPR$1=117")
         .queryContains(
             druidChecker("\"filter\":{\"type\":\"expression\",\"expression\":\""
                     + "(timestamp_floor(timestamp_parse(concat(concat(",
                 "== timestamp_floor(\\\"__time\\\""
-            ))
-        .returnsOrdered("EXPR$0=1997-01-01 00:00:00; EXPR$1=117");
+            ));
   }
 
   @Test
@@ -4382,11 +4388,12 @@ public class DruidAdapterIT {
     final String sql =
         "SELECT CAST(COUNT(*) + SUM(\"store_sales\") as INTEGER) FROM " + FOODMART_TABLE;
     sql(sql, FOODMART)
-        .explainContains("PLAN=EnumerableInterpreter\n"
+        .returnsOrdered("EXPR$0=652067");
+        /*.explainContains("PLAN=EnumerableInterpreter\n"
             + "  BindableProject(EXPR$0=[CAST(+($0, $1)):INTEGER])\n"
             + "    DruidQuery(table=[[foodmart, foodmart]], intervals=[[1900-01-09T00:00:00.000Z/"
-            + "2992-01-10T00:00:00.000Z]], projects=[[$90]], groups=[{}], aggs=[[COUNT(), SUM($0)]])")
-        .returnsOrdered("EXPR$0=652067");
+            + "2992-01-10T00:00:00.000Z]], projects=[[$90]], groups=[{}], aggs=[[COUNT(), SUM($0)]])");*/
+
   }
 
   @Test
